backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2021.5.30
cffi==1.14.6
cfgv==3.2.0
charset-normalizer==2.0.3
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
filelock==3.0.12
gunicorn==20.1.0
identify==2.2.4
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.26
mysqlclient==2.0.3
nodeenv==1.6.0
oauthlib==3.1.1
pathspec==0.8.1
Pillow==8.2.0
pre-commit==2.12.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.26.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.4.4
